<SBRFACCTAX-632> Исправил неправильно передающиеся параметры
git-svn-id: https://horse.aplana.com/svn/SBRFACCTAX/trunk/taxaccounting@548 34aa7802-8fb0-764d-beb8-9b5b9d595954

+++ gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/formdata/client/FormDataPresenter.java
@@ -58,8 +58,8 @@ public class FormDataPresenter extends
 			GetFormData action = new GetFormData();
 			action.setFormDataId(Long.parseLong(request.getParameter(
 					FORM_DATA_ID, String.valueOf(Long.MAX_VALUE))));
-			action.setDepartmentId(Long.parseLong(request.getParameter(
-					DEPARTMENT_ID, String.valueOf(Long.MAX_VALUE))));
+			action.setDepartmentId(Integer.parseInt(request.getParameter(
+					DEPARTMENT_ID, String.valueOf(Integer.MAX_VALUE))));
 			action.setFormDataKind(Long.parseLong((request.getParameter(
 					FORM_DATA_KIND_ID, String.valueOf(Long.MAX_VALUE)))));
 			action.setFormDataTypeId((Long.parseLong((request.getParameter(
